docs: refactor mother items into their own module
